[cleanup] Refactor ContainerDropDescriptionSpec
Extract some methods to reduce the complexity metrics, and move almost
all the code into private static helper methods. The actual custom
metamodel code now simply forwards to these helpers. In a later
phase (as part of #531396), the helpers will be moved outside of the
class, and all callers of the metamodel operations will be changed to
call them directly, so that the operations can then be removed from
the metamodel itself.

Change-Id: I7639793d794e453c2077db5a43b1f85b9872ca80
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>